they're calling it a "woman" moment

+++ code/game/objects/effects/decals/cleanable.dm
@@ -88,6 +88,8 @@
 
 /obj/effect/decal/cleanable/wash(clean_types)
 	..()
+	if(!(flags_1 & INITIALIZED_1))
+		return
 	if(clean_types in list(CLEAN_WASH, CLEAN_SCRUB))
 		qdel(src)
 	return TRUE
